scaling-analytics/requirements.txt
# WattCast v2 — Pre-computation Architecture

# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Message broker
confluent-kafka>=2.3.0

# Cache
redis>=5.0.0

# Database
psycopg2-binary>=2.9.9

# ML (shared with v1 — used by batch scorer)
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.3.0
prophet>=1.1.5
pmdarima>=2.0.0
shap>=0.44.0

# Reused from v1
structlog>=24.1.0
python-dotenv>=1.0.0
holidays>=0.55
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
fakeredis>=2.21.0
